refactor: Repository 레이어 분리
