+++ src/main/java/io/hhplus/arch/infrastructure/LectureRepository.java
@@ -1,5 +1,6 @@
-package io.hhplus.arch.domain.lecture;
+package io.hhplus.arch.infrastructure;
 
+import io.hhplus.arch.domain.lecture.Lecture;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LectureRepository extends JpaRepository<Lecture, Long> {
