+++ src/main/java/io/hhplus/arch/infrastructure/UserRepository.java
@@ -1,5 +1,6 @@
-package io.hhplus.arch.domain.user;
+package io.hhplus.arch.infrastructure;
 
+import io.hhplus.arch.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
